Break the huge promise chain, use a loop
By using a loop over all records instead of building a huge promise chain, we avoid the issue documented in https://github.com/ZJONSSON/node-unzipper/issues/254 where the promise chain ends up taking up all the memory of an application.

In short, this commit allows the garbage collector to collect the small promises as they resolve, rather than building one huge chain that keeps references to the future "then"s.

The downside of this approach currently is that it relies on async/await syntax in order to allow awaits inside loops, which is what can fix the issue here. Im sure we can find a solution without this if it is significant, and we put our minds to it!